service/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.16.1
    # via flask-migrate
aniso8601==10.0.1
    # via flask-restful
bcrypt==4.3.0
    # via -r requirements.in
blinker==1.9.0
    # via flask
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via flask
flask==3.1.1
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-migrate
    #   flask-restful
    #   flask-script
    #   flask-sqlalchemy
flask-cors==6.0.0
    # via -r requirements.in
flask-migrate==4.1.0
    # via -r requirements.in
flask-restful==0.3.10
    # via -r requirements.in
flask-script==2.0.6
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via
    #   -r requirements.in
    #   flask-migrate
gunicorn==23.0.0
    # via -r requirements.in
idna==3.10
    # via requests
imagehash==4.3.2
    # via -r requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   mako
    #   werkzeug
numpy==2.3.0
    # via
    #   imagehash
    #   pywavelets
    #   scipy
packaging==25.0
    # via gunicorn
pillow==11.2.1
    # via
    #   -r requirements.in
    #   imagehash
psycopg2==2.9.10
    # via -r requirements.in
pyjwt==2.10.1
    # via -r requirements.in
pytz==2025.2
    # via flask-restful
pywavelets==1.8.0
    # via
    #   -r requirements.in
    #   imagehash
requests==2.32.3
    # via -r requirements.in
scipy==1.15.3
    # via
    #   -r requirements.in
    #   imagehash
six==1.17.0
    # via flask-restful
sqlalchemy==2.0.41
    # via
    #   -r requirements.in
    #   alembic
    #   flask-sqlalchemy
typing-extensions==4.14.0
    # via
    #   alembic
    #   sqlalchemy
unicodecsv==0.14.1
    # via -r requirements.in
urllib3==2.4.0
    # via requests
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
